+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "qr-order-admin",
   "private": true,
-  "version": "0.3.3",
+  "version": "0.3.4",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -18,7 +18,6 @@
   "dependencies": {
     "@supabase/supabase-js": "^2.46.2",
     "@tanstack/react-query": "^5.61.5",
-    "@tanstack/react-query-devtools": "^5.66.5",
     "dotenv": "^16.4.5",
     "jotai": "^2.12.5",
     "konva": "8.0.1",
@@ -36,6 +35,7 @@
     "@google/genai": "^0.13.0",
     "@playwright/test": "^1.55.0",
     "@tanstack/eslint-plugin-query": "^5.61.4",
+    "@tanstack/react-query-devtools": "^5.66.5",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.0",
     "@types/node": "^24.2.1",
