+++ src/features/modal/tab/menu/components/header/index.module.css
@@ -1,0 +1,40 @@
+.header {
+  width: 100%;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  position: relative;
+}
+
+.modalTitle {
+  color: var(--main-color);
+  font-size: var(--fz-l);
+}
+
+.close {
+  width: 25px;
+  height: 25px;
+
+  display: flex;
+
+  position: absolute;
+  top: 50%;
+  right: 0;
+  box-sizing: content-box;
+
+  transform: translateY(-50%);
+
+  padding: 5px;
+}
+
+.close img {
+  transform: rotateZ(45deg);
+}
+
+@media (max-height: 768px) {
+  .modalTitle {
+    font-size: var(--fz-l-response);
+  }
+}
