+++ src/features/modal/tab/menu/components/image/index.module.css
@@ -1,0 +1,45 @@
+.imgInput {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 15px;
+
+  cursor: pointer;
+}
+
+.imgBox {
+  width: 300px;
+  height: 80px;
+
+  color: var(--sub-color);
+  font-size: var(--fz-s);
+
+  border: 1px solid #d9d9d9;
+  border-radius: 8px;
+  overflow: hidden;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 5px;
+}
+
+.imgBox img {
+  object-fit: contain;
+}
+
+.inputTitle {
+  font-size: var(--fz-r);
+  min-width: 100px;
+  text-align: start;
+}
+
+.iconBox {
+  width: 15px;
+  height: 15px;
+
+  display: flex;
+
+  box-sizing: content-box;
+}
