+++ src/features/modal/tab/menu/index.tsx
@@ -1,9 +1,14 @@
 import { useAtomValue } from 'jotai';
+import { lazy, Suspense } from 'react';
 import { AnimatePresence, motion } from 'motion/react';
 
+import LoadingSpinner from '@/features/load/spinner';
+
+import styles from './index.module.css';
 import { tabModalAtom } from '../store/atom';
-import CreateMenuModal from './create';
-import UpdateMenuModal from './update';
+
+const LazyCreateMenuModal = lazy(() => import('./create'));
+const LazyUpdateMenuModal = lazy(() => import('./update'));
 
 export default function MenuModal() {
   const currentModal = useAtomValue(tabModalAtom);
@@ -13,6 +18,7 @@ export default function MenuModal() {
       {currentModal === 'menu-create' ? (
         <motion.div
           layout
+          className={styles.menuModalBox}
           key={'menu-modal-1'}
           initial={{ x: 385 }}
           animate={{ x: 0 }}
@@ -20,11 +26,14 @@ export default function MenuModal() {
           transition={{ duration: 0.3 }}
           style={{ height: '100%' }}
         >
-          <CreateMenuModal />
+          <Suspense fallback={<LoadingSpinner position='absolute' />}>
+            <LazyCreateMenuModal />
+          </Suspense>
         </motion.div>
       ) : (
         <motion.div
           layout
+          className={styles.menuModalBox}
           key={'menu-modal-2'}
           initial={{ x: 385 }}
           animate={{ x: 0 }}
@@ -32,7 +41,9 @@ export default function MenuModal() {
           transition={{ duration: 0.3 }}
           style={{ height: '100%' }}
         >
-          <UpdateMenuModal />
+          <Suspense fallback={<LoadingSpinner position='absolute' />}>
+            <LazyUpdateMenuModal />
+          </Suspense>
         </motion.div>
       )}
     </AnimatePresence>
