+++ 작업일지.txt
@@ -1411,11 +1411,17 @@ feat: 자동 커밋 및 AI 분석 기능 추가
   - 컴포넌트 개선
     - 이미지 업로드 시 기존 이미지 숨김 처리
 
-<-- 문제 해결 -->
+25.9.12.
+  - 문제 해결
+    - 메뉴 생성/수정 후 메뉴 재생성 시도 
+      - 모달 카테고리 반영 안 됨
+      : 메뉴 사본 상태 초기화 미실행 원인
 
-- 문제
-  - 메뉴 생성 후 메뉴 재생성 시도, 모달 카테고리 반영 안 됨
+      - 파일 미첨부 오류 
 
-  - 메뉴 생성 후 메뉴 재생성 시도, 파일 미첨부 오류 
+  - 메뉴 모달 lazy() 적용 
 
-  - E2E 테스트 필요
+  - E2E 테스트 케이스 수정
+
+<--  -->
+
